Projet BDD/requirements.txt
archan==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
decorator==4.1.2
dependenpy==3.2.0
Django==1.11.5
django-app-settings==0.2.5
django-bootstrap-form==3.3
django-braces==1.11.0
django-cerberus-ac==0.2.0
django-extensions==1.9.6
django-meerkat==0.2.4
django-suit-dashboard==2.0.5
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7

